Add CMake files to build and test using MS-Visual C++.
The CMake files are compatible with other OSs.
Use wild cards for source lists, helps keep CMakeLists.txt in sync
with the project, but forces cmake to be called each time a source
file is added or removed from the list.

Fix RunTests.pl on for MSVC builds

In MSVC builds the executables are stored in PROJECT_BINARY_DIR/Debug or
PROJECT_BINARY_DIR/Release - need to pass the name of the re2c executable
to the perl script RunTests.pl

Accept one test file as input to run just that test

Show diff in case of test failure